==> 2_archive/nowcasting round 2/gas/1_exploration/4_rolling-forecast.R <==
# Rolling forecast

## To do: 
### 1) Check if we need to estimate all ML approaches for each timestamp or can we estimate the
#### models once and then just predict? Need to evaluate.
### 2) Write function for ARIMAX models that estimate the models for each step.
### 3) Perform ensemble learning: Create an approach using ARIMAX and ML combined.
### 4) Estimate for all countries.


# Libs ----

# load data ----

# estimate function 1 ----
